fetching of two pokemons based on property "name" succesful via PokemonFiche component. Moved state to PokemonFiche and only set "name" in function p
